Move disk related types into data module
